production_stack/backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
pydantic==2.10.4
pydantic-settings==2.7.0
python-dotenv==1.0.1
SQLAlchemy==2.0.36
PyMySQL==1.1.1
redis==5.2.1
fakeredis==2.26.2
python-jose==3.3.0
bcrypt>=4.0.1
